// generated/models/ApiModerationCheck.ts
export enum ApiModerationCheckSubjectTypeEnum {
    RepCategory = 'REP_CATEGORY',
    ProfileBio = 'PROFILE_BIO',
    GroupName = 'GROUP_NAME',
    Drop = 'DROP'
}
export enum ApiModerationCheckPolicyFamilyEnum {
    PublicFields = 'PUBLIC_FIELDS',
    WaveContent = 'WAVE_CONTENT'
}
export enum ApiModerationCheckOutcomeEnum {
    Allow = 'ALLOW',
    Reject = 'REJECT',
    Error = 'ERROR',
    Pending = 'PENDING'
}
export enum ApiModerationCheckReviewStatusEnum {
    NeedsReview = 'NEEDS_REVIEW',
    Reviewed = 'REVIEWED'
}